hide the filtered projection traits behind the projection traits
the selection of the base class is done using Kernel::Has_filtered_predicates
