+++ Linear_cell_complex/examples/Linear_cell_complex/linear_cell_complex_3_viewer_qt.h
@@ -32,7 +32,7 @@
 
 #include <CGAL/Linear_cell_complex.h>
 #include <CGAL/Cartesian_converter.h>
-#include <CGAL/Triangulation_2_filtered_projection_traits_3.h>
+#include <CGAL/Triangulation_2_projection_traits_3.h>
 #include <CGAL/Triangulation_vertex_base_with_info_2.h>
 #include <CGAL/Triangulation_face_base_with_info_2.h>
 #include <CGAL/Constrained_Delaunay_triangulation_2.h>
@@ -140,7 +140,7 @@ class SimpleLCCViewerQt : public QGLViewer, public QOpenGLFunctions_2_1
     bool is_process;
   };
 
-  typedef CGAL::Triangulation_2_filtered_projection_traits_3<CGAL::Exact_predicates_inexact_constructions_kernel> P_traits;
+  typedef CGAL::Triangulation_2_projection_traits_3<CGAL::Exact_predicates_inexact_constructions_kernel> P_traits;
   typedef CGAL::Triangulation_vertex_base_with_info_2<Vertex_info, P_traits> Vb;
 
   typedef CGAL::Triangulation_face_base_with_info_2<Face_info, P_traits> Fb1;
