+++ Triangulation_2/include/CGAL/internal/Triangulation_2_filtered_projection_traits_3.h
@@ -22,18 +22,18 @@
 #ifndef CGAL_TRIANGULATION_2_FILTERED_PROJECTION_TRAITS_3_H
 #define CGAL_TRIANGULATION_2_FILTERED_PROJECTION_TRAITS_3_H
 
-#include <CGAL/Triangulation_2_projection_traits_3.h>
+#include <CGAL/internal/Triangulation_2_projection_traits_base_3.h>
 #include <CGAL/Filtered_predicate.h>
 
 namespace CGAL {
 
 template < class Filtered_kernel >
 class Triangulation_2_filtered_projection_traits_3
-  : public Triangulation_2_projection_traits_3<Filtered_kernel>
+  : public Triangulation_2_projection_traits_base_3<Filtered_kernel>
 {
   typedef Filtered_kernel K;
   typedef Triangulation_2_filtered_projection_traits_3<K> Self;
-  typedef Triangulation_2_projection_traits_3<K> Base;
+  typedef Triangulation_2_projection_traits_base_3<K> Base;
 
   typedef typename K::Exact_kernel Exact_kernel;
   typedef typename K::Approximate_kernel Approximate_kernel;
@@ -41,8 +41,8 @@ class Triangulation_2_filtered_projection_traits_3
   typedef typename K::C2F C2F;
 
 public:
-  typedef Triangulation_2_projection_traits_3<Exact_kernel> Exact_traits;
-  typedef Triangulation_2_projection_traits_3<Approximate_kernel> Filtering_traits;
+  typedef Triangulation_2_projection_traits_base_3<Exact_kernel> Exact_traits;
+  typedef Triangulation_2_projection_traits_base_3<Approximate_kernel> Filtering_traits;
 
 public:
   explicit Triangulation_2_filtered_projection_traits_3(const typename K::Vector_3& n)
@@ -77,7 +77,7 @@ public:
   CGAL_TRIANGULATION_2_PROJ_TRAITS_FILTER_PRED(Compare_y_2,
 					       compare_y_2_object,
 					       base2)
-}; // end class Triangulation_2_projection_traits_3<Filtered_kernel>
+}; // end class Triangulation_2_projection_traits_base_3<Filtered_kernel>
 
 } // end namespace CGAL
