+++ Triangulation_2/test/Triangulation_2/test_cdt_2_projection_traits_special_case.cpp
@@ -2,7 +2,7 @@
 #include <CGAL/Exact_predicates_exact_constructions_kernel.h>
 #include <CGAL/Constrained_Delaunay_triangulation_2.h>
 #include <CGAL/Triangulation_2_projection_traits_3.h>
-#include <CGAL/Triangulation_2_filtered_projection_traits_3.h>
+#include <CGAL/Triangulation_2_projection_traits_3.h>
 
 #include <string>
 #include <fstream>
@@ -69,7 +69,7 @@ int main()
   bool ok = true;
   ok = ok &&
       test<CGAL::Epick,
-           CGAL::Triangulation_2_filtered_projection_traits_3<Epick> >
+           CGAL::Triangulation_2_projection_traits_3<Epick> >
       ("CDT_2 in a 3D plane, with Epick");
   ok = ok &&
       test<CGAL::Epeck,
